fantasy_trade_calculator_deployment/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
numpy>=1.26.0
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0
pytz==2023.3.post1
python-dateutil==2.8.2
six==1.16.0
tzdata==2023.3
et-xmlfile==1.1.0
packaging==23.2
psycopg2-binary>=2.9.10
Flask-Caching==2.1.0
SQLAlchemy>=2.0.36